+++ tools/vp_visualize_base.py
@@ -1,0 +1,72 @@
+import numpy as np
+import open3d as o3d
+from scipy.spatial.transform import Rotation
+from scipy.spatial.transform import Rotation as R
+
+def visualize_selected_viewpoints(object_mesh, viewpoint_6Ds_list_base, frustum_size=0.1):
+    """
+    Visualize selected viewpoints as coordinate frames along with the object mesh.
+    
+    Parameters:
+        object_mesh: The 3D mesh of the object (in base frame).
+        viewpoint_6Ds_list_base: List of viewpoint poses in base frame as [x, y, z, ox, oy, oz, ow].
+        frustum_size: Scaling factor for the camera frustum visualization.
+    """
+
+    # Load object mesh if it's a file path
+    if isinstance(object_mesh, str):
+        object_mesh = o3d.io.read_triangle_mesh(object_mesh)
+    
+    visualization_objects = [object_mesh]
+
+    # Iterate over each viewpoint pose
+    for vp in viewpoint_6Ds_list_base:
+        x, y, z, ox, oy, oz, ow = vp
+
+        # Convert quaternion to rotation matrix
+        rotation_matrix = R.from_quat([ox, oy, oz, ow]).as_matrix()
+
+        # Create a coordinate frame for the viewpoint
+        frame = o3d.geometry.TriangleMesh.create_coordinate_frame(size=0.05, origin=[x, y, z])
+        frame.rotate(rotation_matrix, center=[x, y, z])
+        visualization_objects.append(frame)
+
+    origin_frame = o3d.geometry.TriangleMesh.create_coordinate_frame(size=0.2)
+    visualization_objects.append(origin_frame)
+
+    # Create a custom visualization window
+    vis = o3d.visualization.Visualizer()
+    vis.create_window("filtered_vp_visualize_base")  # Set custom window title
+
+    for geom in visualization_objects:
+        vis.add_geometry(geom)
+
+    vis.run()
+    vis.destroy_window()
+
+
+file_path = "data/8400310XKM42A_new.STL"  # Replace with your CAD model path
+
+# Given transformation from base to object
+base_to_object_translation = np.array([1.38494077,-0.07737339,0.95311693])  # (x, y, z)
+base_to_object_rpy = np.array([0.00784205,0.016702,1.64769838])  # (roll, pitch, yaw)
+
+# Convert RPY to rotation matrix
+base_to_object_rotation = Rotation.from_euler('xyz', base_to_object_rpy).as_matrix()
+
+# Construct homogeneous transformation matrix (4x4)
+T_base_object = np.eye(4)
+T_base_object[:3, :3] = base_to_object_rotation
+T_base_object[:3, 3] = base_to_object_translation
+
+# Load object STL (in object frame)
+object_mesh = o3d.io.read_triangle_mesh(file_path)
+
+# Apply transformation to move STL to base frame
+object_mesh.transform(T_base_object)
+
+viewpoint_6Ds_list_base = [[0.8376502379637294, -0.35666256777008454, 0.8941516665888057, 0.6930584744288885, 0.6983073374106553, -0.018991976902317055, -0.17797785916816963], [0.41892496070527335, -0.15711894631368498, 0.07974913385643856, 0.25731253357701156, 0.08342666740794778, 0.6663117165760516, 0.6948805275605718], [0.4926366952378988, 0.378208617569304, 0.784015404196923, 0.6766618181224053, 0.656347981364828, -0.009200515155022976, 0.3335437928885479], [0.9369727416936172, -0.7071710339208248, 0.6776364121389593, 0.7043317097203702, 0.5364943234155035, 0.03812949207870314, -0.4632891380759336], [1.0321238353385236, -0.015223902443111004, 0.1479615716774343, -0.18272989484955385, -0.24948081024731003, 0.7148981231529284, 0.6271282040851275], [0.14628046044132845, -0.2552391489845509, 0.34814697718542664, 0.43458360702931387, 0.434632216305034, 0.5336313104334227, 0.5809212937879563], [0.9764118008964461, -0.3242785106888701, 0.9515668344404312, 0.6915905702166971, 0.6890056148638007, -0.08922413448654085, -0.19751658081950346], [0.7114138818146404, -0.36237895768657724, 0.05294318751092386, -0.057052162252784415, 0.07176130264524345, 0.7250742175098749, 0.682541387263353], [0.3617698629100631, 0.5433958214583245, 0.6727182350154683, 0.6740140806595625, 0.5749623068449784, 0.02298670246711411, 0.46324397059199196], [0.7890691405608307, -0.6603469839705285, 0.03332900637816427, -0.2537723197553523, 0.05911950443245152, 0.6881316401308519, 0.6771848638093385], [0.39116053067641965, 0.08068420982818703, 0.9253724515068787, 0.6757715004189762, 0.6952586911426829, 0.04789749394797596, 0.2401126020965016], [0.18124765408265558, -0.47915732469067684, 0.5641051813761694, 0.4905887239123103, 0.6405481954731091, 0.46640573489287324, 0.36261053998094633], [0.5497774627628581, -0.5805584190633354, 0.8104861834322541, 0.6090348407092459, 0.7312914099291307, 0.29835527881528145, -0.07261930989195317], [0.5438794154375973, -0.47958697001466655, 0.8235238026459388, 0.6271062161951331, 0.7353766591142311, 0.2568212104905485, 0.001352287537714353], [0.9527993480440776, -0.26710422725278415, 0.9276178598514264, 0.6886438907751282, 0.691550089992562, -0.11334594321781445, -0.1862277151377767], [0.6555373034742435, 0.01782148291329773, 0.9488926464429724, 0.6833940037857948, 0.7251265842360174, -0.03148585975755561, 0.07856661542382168]]
+
+
+filtered_viewpoint_6Ds_list_base= [[0.6867226711642969, -0.16735937208524337, 1.0505542339192298, 0.6788538285444745, 0.7339434948519156, 0.021965476329240643, 0.0013941609287822928], [0.5616927021124545, 0.2756731096990191, 0.9382356505932723, 0.6785549897523498, 0.702511631913084, 0.022631032239113134, 0.21337377835337829], [0.971218352098181, -0.6593979622126203, 0.8086198994987587, 0.6795186436579507, 0.6316612284862883, 0.0791801470478401, -0.36467658227796135], [0.09892243689217967, -0.3514702692923795, 0.4170129374689871, 0.5199867602567818, 0.4864051912219252, 0.4697239618606543, 0.5219033998394234], [0.527664963076638, 0.49602348778595623, 0.8431451362546963, 0.6767386718493114, 0.6704142204583059, -0.054980489716510164, 0.2992435275514663], [0.9549013073314224, -0.9104144627101289, 0.44321857248525864, 0.6951538286652886, 0.36654027967304476, 0.04409629978225637, -0.6168183640381585], [0.2375409398387873, 0.2082496121648239, 0.7836049848662923, 0.6473358586180321, 0.6483974236615289, 0.2196958737245548, 0.335068336619191], [0.6937687850399258, -0.9548529083597126, 0.05296729090918323, -0.3265779358954537, 0.4576912037899682, 0.654301852918589, 0.5057219582209735], [0.9094608343153788, -0.3958974808476241, 0.983976222852183, 0.6826139208507719, 0.6969220697365749, -0.0043114449421593985, -0.21981645802199537], [0.24417268995558056, -0.7261619650118165, 0.6600190572350838, 0.4861014096117179, 0.7050000911681036, 0.4864552444366765, 0.17332508816244738], [0.6858422153427592, -0.6541875906297192, 0.9283965443196991, 0.6606491456411647, 0.7086105782243077, 0.15922359667034292, -0.1898989232606953], [0.7891327543630978, 0.10944158609021573, 1.0199576375385226, 0.6756074889992415, 0.7316007323463618, -0.0739377377364286, 0.053367594819573705], [0.6037065467269511, -0.9527261548394581, 0.6558569539704882, 0.40505317172805144, 0.6925664866027348, 0.5472994623206268, -0.2382160537229648], [0.5299992704093025, -0.4908614907430492, 0.9911695777873255, 0.6426803109278487, 0.7300304204953989, 0.22039339907209288, 0.07378585733353141], [1.0688024496484971, -0.646518003618876, 0.7603383945023204, 0.6887191025081496, 0.5848940847572791, -0.025627716087791044, -0.4276775977584488], [0.38837536472660095, -0.7054444036634059, 0.8248381874850188, 0.5086545307182608, 0.7302690569829668, 0.4556255129160635, 0.019572040592850084], [0.2850207646060625, 0.08820982301489547, 0.9068028832223554, 0.6531985469160391, 0.692923084273453, 0.19272643417864027, 0.23673989768631482], [0.3383225734777639, 0.06968151364912266, 0.9471749647656695, 0.6638676912611583, 0.7029734215649259, 0.14707790616329275, 0.2085093441347466], [0.47239904815901756, -0.5742043745526186, 0.9527484946605373, 0.6496917762813588, 0.72993837512687, 0.19816690504577145, 0.07629182193458263], [0.42175484440182615, -0.6460058863773279, 0.8904730385933409, 0.6250449969066628, 0.7295929367859694, 0.26767821197087405, 0.07322071612631409], [1.1551475190956024, -0.717665317723827, 0.07379955501733976, -0.4659306401038632, -0.28157043100779744, 0.5542733093136409, 0.6296092673849334], [1.0878242671776963, -0.428688455613508, 0.9027871951144187, 0.679496297265807, 0.6703506165118178, -0.09809245237008787, -0.2815896016054323], [0.5354750886748572, 0.629406023120572, 0.7669086274086581, 0.6654103092940918, 0.6393760910255608, -0.1346115464247004, 0.3609807004247813], [0.7315748753334168, -0.6109513780093168, 0.951102387597562, 0.6525354270289786, 0.7236472811842066, 0.18736422654233773, -0.12420457124842882], [0.677930300361497, -0.5378392510649523, 0.9861889487009351, 0.6500108163004346, 0.7284594884863335, 0.19551704478168558, -0.09276743785676486], [0.38774555620465495, 0.2856010090850077, 0.8641240769277904, 0.6746421565605119, 0.6724778495672306, 0.0870593912521366, 0.29163704296200543], [0.6407481561105935, -0.36581902809051053, 1.0277357155772255, 0.669355856190185, 0.729257493405159, 0.11466561089401556, -0.08365431115388598]]
+visualize_selected_viewpoints(object_mesh, filtered_viewpoint_6Ds_list_base)
